+++ crates/core/query/src/lib.rs
@@ -1,4 +1,7 @@
+mod builder;
 mod errors;
 mod parser;
 
+pub use builder::*;
+pub use errors::QueryError;
 pub use parser::*;
